drop unused PoCIndelCaller
fold PoCIndelCaller’s tests into SomaticStandard’s

fixes #207
